+++ class/Probe.h
@@ -36,8 +36,8 @@ public:
   size_t DataRate() const	{ return(_dataRate); }
   float FillValue() const	{ return _missing_value; }
 
-  short	FirstBin() const	{ return(_firstBin); }
-  short	LastBin() const		{ return(_lastBin); }
+  size_t FirstBin() const	{ return(_firstBin); }
+  size_t LastBin() const	{ return(_lastBin); }
 
   bool	HaveConcentrations() const	{ return(_cvar); }
